Move FileTypeClassifier from io_utils to classifiers package
