+++ haystack/preview/components/__init__.py
@@ -1,4 +1,4 @@
 from haystack.preview.components.audio.whisper_local import LocalWhisperTranscriber
 from haystack.preview.components.audio.whisper_remote import RemoteWhisperTranscriber
 from haystack.preview.components.file_converters import TextFileToDocument
-from haystack.preview.components.io_utils import FileTypeClassifier
+from haystack.preview.components.classifiers import FileTypeClassifier
