+++ haystack/preview/components/classifiers/__init__.py
@@ -1,0 +1,1 @@
+from haystack.preview.components.classifiers.file_classifier import FileTypeClassifier
