--- haystack/preview/components/io_utils/__init__.py
@@ -1,1 +1,0 @@
-from haystack.preview.components.io_utils.file_classifier import FileTypeClassifier
